Create CombinesFilterbank
Combines a seris of filterbank files together and plots them in a frequency-time waterfall graph. Sorts all the filterbank files(without headers) in a list, takes the first 131072 data points of the first spectrum, appends it to the combined file, takes the next 131072 data points and appends it, and so on until it was one spectrum after another, then added a header to that big combined file.